merge

+++ cluster-autoscaler/core/utils.go
@@ -18,23 +18,27 @@ package core
 
 import (
 	"fmt"
+	"math"
 	"math/rand"
 	"reflect"
 	"time"
 
 	"k8s.io/autoscaler/cluster-autoscaler/cloudprovider"
 	"k8s.io/autoscaler/cluster-autoscaler/clusterstate"
+	"k8s.io/autoscaler/cluster-autoscaler/clusterstate/utils"
+	"k8s.io/autoscaler/cluster-autoscaler/metrics"
 	"k8s.io/autoscaler/cluster-autoscaler/simulator"
 	"k8s.io/autoscaler/cluster-autoscaler/utils/daemonset"
+	"k8s.io/autoscaler/cluster-autoscaler/utils/deletetaint"
+	"k8s.io/autoscaler/cluster-autoscaler/utils/drain"
 	"k8s.io/autoscaler/cluster-autoscaler/utils/errors"
 	kube_util "k8s.io/autoscaler/cluster-autoscaler/utils/kubernetes"
+	scheduler_util "k8s.io/autoscaler/cluster-autoscaler/utils/scheduler"
 
 	apiv1 "k8s.io/api/core/v1"
 	extensionsv1 "k8s.io/api/extensions/v1beta1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	api "k8s.io/kubernetes/pkg/api"
-	podv1 "k8s.io/kubernetes/pkg/api/v1/pod"
-	kube_client "k8s.io/kubernetes/pkg/client/clientset_generated/clientset"
+	apiequality "k8s.io/apimachinery/pkg/api/equality"
+	kube_client "k8s.io/client-go/kubernetes"
 	kubeletapis "k8s.io/kubernetes/pkg/kubelet/apis"
 	"k8s.io/kubernetes/plugin/pkg/scheduler/schedulercache"
 
@@ -46,98 +50,121 @@ const (
 	ReschedulerTaintKey = "CriticalAddonsOnly"
 )
 
-// GetAllNodesAvailableTime returns time when the newest node became available for scheduler.
-// TODO: This function should use LastTransitionTime from NodeReady condition.
-func GetAllNodesAvailableTime(nodes []*apiv1.Node) time.Time {
-	var result time.Time
-	for _, node := range nodes {
-		if node.CreationTimestamp.After(result) {
-			result = node.CreationTimestamp.Time
-		}
-	}
-	return result.Add(1 * time.Minute)
+// Following data structure is used to avoid running predicates #pending_pods * #nodes
+// times (which turned out to be very expensive if there are thousands of pending pods).
+// This optimization is based on the assumption that if there are that many pods they're
+// likely created by controllers (deployment, replication controller, ...).
+// So instead of running all predicates for every pod we first check whether we've
+// already seen identical pod (in this step we're not binpacking, just checking if
+// the pod would fit anywhere right now) and if so we use the result we already
+// calculated.
+// To decide if two pods are similar enough we check if they have identical label
+// and spec and are owned by the same controller. The problem is the whole
+// podSchedulableInfo struct is not hashable and keeping a list and runnig deep
+// equality checks would likely also be expensive. So instead we use controller
+// UID as a key in initial lookup and only run full comparison on a set of
+// podSchedulableInfos created for pods owned by this controller.
+type podSchedulableInfo struct {
+	spec        apiv1.PodSpec
+	labels      map[string]string
+	schedulable bool
 }
 
-// SlicePodsByPodScheduledTime slices given pod array into those where PodScheduled condition
-// have been updated after the thresold and others.
-// Each pod must be in condition "Scheduled: False; Reason: Unschedulable"
-func SlicePodsByPodScheduledTime(pods []*apiv1.Pod, threshold time.Time) (oldPods []*apiv1.Pod, newPods []*apiv1.Pod) {
-	for _, pod := range pods {
-		_, condition := podv1.GetPodCondition(&pod.Status, apiv1.PodScheduled)
-		if condition != nil {
-			if condition.LastTransitionTime.After(threshold) {
-				newPods = append(newPods, pod)
-			} else {
-				oldPods = append(oldPods, pod)
+type podSchedulableMap map[string][]podSchedulableInfo
+
+func (psi *podSchedulableInfo) match(pod *apiv1.Pod) bool {
+	return reflect.DeepEqual(pod.Labels, psi.labels) && apiequality.Semantic.DeepEqual(pod.Spec, psi.spec)
+}
+
+func (podMap podSchedulableMap) get(pod *apiv1.Pod) (bool, bool) {
+	ref := drain.ControllerRef(pod)
+	if ref == nil {
+		return false, false
+	}
+	uid := string(ref.UID)
+	if infos, found := podMap[uid]; found {
+		for _, info := range infos {
+			if info.match(pod) {
+				return info.schedulable, true
 			}
 		}
 	}
-	return
+	return false, false
 }
 
-// ResetPodScheduledCondition resets pod condition PodScheduled to "unknown" for all the pods with LastTransitionTime
-// not after the threshold time.
-func ResetPodScheduledCondition(kubeClient kube_client.Interface, pods []*apiv1.Pod) {
-	for _, pod := range pods {
-		if err := resetPodScheduledConditionForPod(kubeClient, pod); err != nil {
-			glog.Errorf("Error during reseting pod condition for %s/%s: %v", pod.Namespace, pod.Name, err)
-		}
+func (podMap podSchedulableMap) set(pod *apiv1.Pod, schedulable bool) {
+	ref := drain.ControllerRef(pod)
+	if ref == nil {
+		return
 	}
-}
-
-func resetPodScheduledConditionForPod(kubeClient kube_client.Interface, pod *apiv1.Pod) error {
-	_, condition := podv1.GetPodCondition(&pod.Status, apiv1.PodScheduled)
-	if condition != nil {
-		glog.V(4).Infof("Reseting pod condition for %s/%s, last transition: %s",
-			pod.Namespace, pod.Name, condition.LastTransitionTime.Time.String())
-		condition.Status = apiv1.ConditionUnknown
-		condition.LastTransitionTime = metav1.Time{
-			Time: time.Now(),
-		}
-		_, err := kubeClient.Core().Pods(pod.Namespace).UpdateStatus(pod)
-		return err
-	}
-	return fmt.Errorf("Expected condition PodScheduled")
+	uid := string(ref.UID)
+	podMap[uid] = append(podMap[uid], podSchedulableInfo{
+		spec:        pod.Spec,
+		labels:      pod.Labels,
+		schedulable: schedulable,
+	})
 }
 
 // FilterOutSchedulable checks whether pods from <unschedulableCandidates> marked as unschedulable
 // by Scheduler actually can't be scheduled on any node and filter out the ones that can.
-func FilterOutSchedulable(unschedulableCandidates []*apiv1.Pod, nodes []*apiv1.Node, allPods []*apiv1.Pod, predicateChecker *simulator.PredicateChecker) []*apiv1.Pod {
+// It takes into account pods that are bound to node and will be scheduled after lower priority pod preemption.
+func FilterOutSchedulable(unschedulableCandidates []*apiv1.Pod, nodes []*apiv1.Node, allScheduled []*apiv1.Pod, podsWaitingForLowerPriorityPreemption []*apiv1.Pod,
+	predicateChecker *simulator.PredicateChecker, expendablePodsPriorityCutoff int) []*apiv1.Pod {
+
 	unschedulablePods := []*apiv1.Pod{}
-	nodeNameToNodeInfo := createNodeNameToInfoMap(allPods, nodes)
+	nonExpendableScheduled := FilterOutExpendablePods(allScheduled, expendablePodsPriorityCutoff)
+	nodeNameToNodeInfo := scheduler_util.CreateNodeNameToInfoMap(append(nonExpendableScheduled, podsWaitingForLowerPriorityPreemption...), nodes)
+	podSchedulable := make(podSchedulableMap)
 
 	for _, pod := range unschedulableCandidates {
+		if schedulable, found := podSchedulable.get(pod); found {
+			if !schedulable {
+				unschedulablePods = append(unschedulablePods, pod)
+			} else {
+				glog.V(4).Infof("Pod %s marked as unschedulable can be scheduled (based on simulation run for other pod owned by the same controller). Ignoring in scale up.", pod.Name)
+			}
+			continue
+		}
 		if nodeName, err := predicateChecker.FitsAny(pod, nodeNameToNodeInfo); err == nil {
-			glog.Warningf("Pod %s marked as unschedulable can be scheduled on %s. Ignoring in scale up.", pod.Name, nodeName)
+			glog.V(4).Infof("Pod %s marked as unschedulable can be scheduled on %s. Ignoring in scale up.", pod.Name, nodeName)
+			podSchedulable.set(pod, true)
 		} else {
 			unschedulablePods = append(unschedulablePods, pod)
+			podSchedulable.set(pod, false)
 		}
 	}
 
 	return unschedulablePods
 }
 
-// TODO: move this function to scheduler utils.
-func createNodeNameToInfoMap(pods []*apiv1.Pod, nodes []*apiv1.Node) map[string]*schedulercache.NodeInfo {
-	nodeNameToNodeInfo := schedulercache.CreateNodeNameToInfoMap(pods, nodes)
-	for _, node := range nodes {
-		if nodeInfo, found := nodeNameToNodeInfo[node.Name]; found {
-			nodeInfo.SetNode(node)
+// FilterOutExpendableAndSplit filters out expendable pods and splits into:
+//   - waiting for lower priority pods preemption
+//   - other pods.
+func FilterOutExpendableAndSplit(unschedulableCandidates []*apiv1.Pod, expendablePodsPriorityCutoff int) ([]*apiv1.Pod, []*apiv1.Pod) {
+	unschedulableNonExpendable := []*apiv1.Pod{}
+	waitingForLowerPriorityPreemption := []*apiv1.Pod{}
+	for _, pod := range unschedulableCandidates {
+		if pod.Spec.Priority != nil && int(*pod.Spec.Priority) < expendablePodsPriorityCutoff {
+			glog.V(4).Infof("Pod %s has priority below %d (%d) and will scheduled when enough resources is free. Ignoring in scale up.", pod.Name, expendablePodsPriorityCutoff, *pod.Spec.Priority)
+		} else if annot, found := pod.Annotations[scheduler_util.NominatedNodeAnnotationKey]; found && len(annot) > 0 {
+			waitingForLowerPriorityPreemption = append(waitingForLowerPriorityPreemption, pod)
+			glog.V(4).Infof("Pod %s will be scheduled after low prioity pods are preempted on %s. Ignoring in scale up.", pod.Name, annot)
+		} else {
+			unschedulableNonExpendable = append(unschedulableNonExpendable, pod)
 		}
 	}
+	return unschedulableNonExpendable, waitingForLowerPriorityPreemption
+}
 
-	// Some pods may be out of sync with node lists. Removing incomplete node infos.
-	keysToRemove := make([]string, 0)
-	for key, nodeInfo := range nodeNameToNodeInfo {
-		if nodeInfo.Node() == nil {
-			keysToRemove = append(keysToRemove, key)
+// FilterOutExpendablePods filters out expendable pods.
+func FilterOutExpendablePods(pods []*apiv1.Pod, expendablePodsPriorityCutoff int) []*apiv1.Pod {
+	result := []*apiv1.Pod{}
+	for _, pod := range pods {
+		if pod.Spec.Priority == nil || int(*pod.Spec.Priority) >= expendablePodsPriorityCutoff {
+			result = append(result, pod)
 		}
 	}
-	for _, key := range keysToRemove {
-		delete(nodeNameToNodeInfo, key)
-	}
-
-	return nodeNameToNodeInfo
+	return result
 }
 
 // GetNodeInfosForGroups finds NodeInfos for all node groups used to manage the given nodes. It also returns a node group to sample node mapping.
@@ -245,11 +272,7 @@ func sanitizeNodeInfo(nodeInfo *schedulercache.NodeInfo, nodeGroupName string) (
 	// Update nodename in pods.
 	sanitizedPods := make([]*apiv1.Pod, 0)
 	for _, pod := range nodeInfo.Pods() {
-		obj, err := api.Scheme.DeepCopy(pod)
-		if err != nil {
-			return nil, errors.ToAutoscalerError(errors.InternalError, err)
-		}
-		sanitizedPod := obj.(*apiv1.Pod)
+		sanitizedPod := pod.DeepCopy()
 		sanitizedPod.Spec.NodeName = sanitizedNode.Name
 		sanitizedPods = append(sanitizedPods, sanitizedPod)
 	}
@@ -263,12 +286,8 @@ func sanitizeNodeInfo(nodeInfo *schedulercache.NodeInfo, nodeGroupName string) (
 }
 
 func sanitizeTemplateNode(node *apiv1.Node, nodeGroup string) (*apiv1.Node, errors.AutoscalerError) {
-	obj, err := api.Scheme.DeepCopy(node)
-	if err != nil {
-		return nil, errors.ToAutoscalerError(errors.InternalError, err)
-	}
+	newNode := node.DeepCopy()
 	nodeName := fmt.Sprintf("template-node-for-%s-%d", nodeGroup, rand.Int63())
-	newNode := obj.(*apiv1.Node)
 	newNode.Labels = make(map[string]string, len(node.Labels))
 	for k, v := range node.Labels {
 		if k != kubeletapis.LabelHostname {
@@ -283,9 +302,12 @@ func sanitizeTemplateNode(node *apiv1.Node, nodeGroup string) (*apiv1.Node, erro
 		// Rescheduler can put this taint on a node while evicting non-critical pods.
 		// New nodes will not have this taint and so we should strip it when creating
 		// template node.
-		if taint.Key == ReschedulerTaintKey {
+		switch taint.Key {
+		case ReschedulerTaintKey:
 			glog.V(4).Infof("Removing rescheduler taint when creating template from node %s", node.Name)
-		} else {
+		case deletetaint.ToBeDeletedTaint:
+			glog.V(4).Infof("Removing autoscaler taint when creating template from node %s", node.Name)
+		default:
 			newTaints = append(newTaints, taint)
 		}
 	}
@@ -293,12 +315,12 @@ func sanitizeTemplateNode(node *apiv1.Node, nodeGroup string) (*apiv1.Node, erro
 	return newNode, nil
 }
 
-// Removes unregisterd nodes if needed. Returns true if anything was removed and error if such occurred.
+// Removes unregistered nodes if needed. Returns true if anything was removed and error if such occurred.
 func removeOldUnregisteredNodes(unregisteredNodes []clusterstate.UnregisteredNode, context *AutoscalingContext,
-	currentTime time.Time) (bool, error) {
+	currentTime time.Time, logRecorder *utils.LogEventRecorder) (bool, error) {
 	removedAny := false
 	for _, unregisteredNode := range unregisteredNodes {
-		if unregisteredNode.UnregisteredSince.Add(context.UnregisteredNodeRemovalTime).Before(currentTime) {
+		if unregisteredNode.UnregisteredSince.Add(context.MaxNodeProvisionTime).Before(currentTime) {
 			glog.V(0).Infof("Removing unregistered node %v", unregisteredNode.Node.Name)
 			nodeGroup, err := context.CloudProvider.NodeGroupForNode(unregisteredNode.Node)
 			if err != nil {
@@ -309,6 +331,17 @@ func removeOldUnregisteredNodes(unregisteredNodes []clusterstate.UnregisteredNod
 				glog.Warningf("No node group for node %s, skipping", unregisteredNode.Node.Name)
 				continue
 			}
+			size, err := nodeGroup.TargetSize()
+			if err != nil {
+				glog.Warningf("Failed to get node group size, err: %v", err)
+				continue
+			}
+			if nodeGroup.MinSize() >= size {
+				glog.Warningf("Failed to remove node %s: node group min size reached, skipping unregistered node removal", unregisteredNode.Node.Name)
+				continue
+			}
+			logRecorder.Eventf(apiv1.EventTypeNormal, "DeleteUnregistered",
+				"Removing unregistered node %v", unregisteredNode.Node.Name)
 			err = nodeGroup.DeleteNodes([]*apiv1.Node{unregisteredNode.Node})
 			if err != nil {
 				glog.Warningf("Failed to remove node %s: %v", unregisteredNode.Node.Name, err)
@@ -330,7 +363,7 @@ func fixNodeGroupSize(context *AutoscalingContext, currentTime time.Time) (bool,
 		if incorrectSize == nil {
 			continue
 		}
-		if incorrectSize.FirstObserved.Add(context.UnregisteredNodeRemovalTime).Before(currentTime) {
+		if incorrectSize.FirstObserved.Add(context.MaxNodeProvisionTime).Before(currentTime) {
 			delta := incorrectSize.CurrentSize - incorrectSize.ExpectedSize
 			if delta < 0 {
 				glog.V(0).Infof("Decreasing size of %s, expected=%d current=%d delta=%d", nodeGroup.Id(),
@@ -347,9 +380,14 @@ func fixNodeGroupSize(context *AutoscalingContext, currentTime time.Time) (bool,
 	return fixed, nil
 }
 
-// getManagedNodes returns the nodes managed by the cluster autoscaler.
-func getManagedNodes(context *AutoscalingContext, nodes []*apiv1.Node) []*apiv1.Node {
+// getPotentiallyUnneededNodes returns nodes that are:
+// - managed by the cluster autoscaler
+// - in groups with size > min size
+func getPotentiallyUnneededNodes(context *AutoscalingContext, nodes []*apiv1.Node) []*apiv1.Node {
 	result := make([]*apiv1.Node, 0, len(nodes))
+
+	nodeGroupSize := getNodeGroupSizeMap(context.CloudProvider)
+
 	for _, node := range nodes {
 		nodeGroup, err := context.CloudProvider.NodeGroupForNode(node)
 		if err != nil {
@@ -360,7 +398,128 @@ func getManagedNodes(context *AutoscalingContext, nodes []*apiv1.Node) []*apiv1.
 			glog.V(4).Infof("Skipping %s - no node group config", node.Name)
 			continue
 		}
+		size, found := nodeGroupSize[nodeGroup.Id()]
+		if !found {
+			glog.Errorf("Error while checking node group size %s: group size not found", nodeGroup.Id())
+			continue
+		}
+		if size <= nodeGroup.MinSize() {
+			glog.V(1).Infof("Skipping %s - node group min size reached", node.Name)
+			continue
+		}
 		result = append(result, node)
 	}
 	return result
+}
+
+func hasHardInterPodAffinity(affinity *apiv1.Affinity) bool {
+	if affinity == nil {
+		return false
+	}
+	if affinity.PodAffinity != nil {
+		if len(affinity.PodAffinity.RequiredDuringSchedulingIgnoredDuringExecution) > 0 {
+			return true
+		}
+	}
+	if affinity.PodAntiAffinity != nil {
+		if len(affinity.PodAntiAffinity.RequiredDuringSchedulingIgnoredDuringExecution) > 0 {
+			return true
+		}
+	}
+	return false
+}
+
+func anyPodHasHardInterPodAffinity(pods []*apiv1.Pod) bool {
+	for _, pod := range pods {
+		if hasHardInterPodAffinity(pod.Spec.Affinity) {
+			return true
+		}
+	}
+	return false
+}
+
+// ConfigurePredicateCheckerForLoop can be run to update predicateChecker configuration
+// based on current state of the cluster.
+func ConfigurePredicateCheckerForLoop(unschedulablePods []*apiv1.Pod, schedulablePods []*apiv1.Pod, predicateChecker *simulator.PredicateChecker) {
+	podsWithAffinityFound := anyPodHasHardInterPodAffinity(unschedulablePods)
+	if !podsWithAffinityFound {
+		podsWithAffinityFound = anyPodHasHardInterPodAffinity(schedulablePods)
+	}
+	predicateChecker.SetAffinityPredicateEnabled(podsWithAffinityFound)
+	if !podsWithAffinityFound {
+		glog.V(1).Info("No pod using affinity / antiaffinity found in cluster, disabling affinity predicate for this loop")
+	}
+}
+
+// Getting node cores/memory
+const (
+	// Megabyte is 2^20 bytes.
+	Megabyte float64 = 1024 * 1024
+)
+
+func getNodeCoresAndMemory(node *apiv1.Node) (int64, int64, error) {
+	cores, err := getNodeResource(node, apiv1.ResourceCPU)
+	if err != nil {
+		return 0, 0, err
+	}
+
+	memory, err := getNodeResource(node, apiv1.ResourceMemory)
+	if err != nil {
+		return 0, 0, err
+	}
+
+	if cores <= 0 || memory <= 0 {
+		return 0, 0, fmt.Errorf("Invalid node CPU/memory values - cpu %v, memory %v", cores, memory)
+	}
+
+	memoryMb := math.Ceil(float64(memory) / Megabyte)
+	return cores, int64(memoryMb), nil
+}
+
+func getNodeResource(node *apiv1.Node, resource apiv1.ResourceName) (int64, error) {
+	nodeCapacity, found := node.Status.Capacity[resource]
+	if !found {
+		return 0, fmt.Errorf("Failed to get %v for node %v", resource, node.Name)
+	}
+	return nodeCapacity.Value(), nil
+}
+
+func getNodeGroupSizeMap(cloudProvider cloudprovider.CloudProvider) map[string]int {
+	nodeGroupSize := make(map[string]int)
+	for _, nodeGroup := range cloudProvider.NodeGroups() {
+		size, err := nodeGroup.TargetSize()
+		if err != nil {
+			glog.Errorf("Error while checking node group size %s: %v", nodeGroup.Id(), err)
+			continue
+		}
+		nodeGroupSize[nodeGroup.Id()] = size
+	}
+	return nodeGroupSize
+}
+
+// UpdateClusterStateMetrics updates metrics related to cluster state
+func UpdateClusterStateMetrics(csr *clusterstate.ClusterStateRegistry) {
+	if csr == nil || reflect.ValueOf(csr).IsNil() {
+		return
+	}
+	metrics.UpdateClusterSafeToAutoscale(csr.IsClusterHealthy())
+	readiness := csr.GetClusterReadiness()
+	metrics.UpdateNodesCount(readiness.Ready, readiness.Unready+readiness.LongNotStarted, readiness.NotStarted, readiness.LongUnregistered, readiness.Unregistered)
+}
+
+func getOldestCreateTime(pods []*apiv1.Pod) time.Time {
+	oldest := time.Now()
+	for _, pod := range pods {
+		if oldest.After(pod.CreationTimestamp.Time) {
+			oldest = pod.CreationTimestamp.Time
+		}
+	}
+	return oldest
+}
+
+// UpdateEmptyClusterStateMetrics updates metrics related to empty cluster's state.
+// TODO(aleksandra-malinowska): use long unregistered value from ClusterStateRegistry.
+func UpdateEmptyClusterStateMetrics() {
+	metrics.UpdateClusterSafeToAutoscale(false)
+	metrics.UpdateNodesCount(0, 0, 0, 0, 0)
 }
